+++ wavenet/audio_reader.py
@@ -66,9 +66,9 @@ class AudioReader(object):
         self.sample_size = sample_size
         self.silence_threshold = silence_threshold
         self.threads = []
-        self.sample_placeholder = tf.placeholder(dtype=tf.float32, shape=None)
+        self.sample_placeholder = tf.placeholder(dtype=tf.float16, shape=None)
         self.queue = tf.PaddingFIFOQueue(queue_size,
-                                         ['float32'],
+                                         ['float16'],
                                          shapes=[(None, 1)])
         self.enqueue = self.queue.enqueue([self.sample_placeholder])
